# specificatie/gba-openapi.yaml
openapi: 3.0.3
servers:
  - description: "APILAB testserver historie"
    url: https://www.haalcentraal.nl/haalcentraal/api/brphistorie
info:
  title: BRP historie bevragen
  description: |
                API voor het zoeken en raadplegen van historische verblijfplaatsen, partners, nationaliteiten en verblijfstitels uit de BRP (inclusief de RNI).

                Zie de [Functionele documentatie](https://github.com/VNG-Realisatie/Haal-Centraal-BRP-historie-bevragen/tree/v1.0.0/features) voor nadere toelichting.
  version: 2.0.0
  contact:
    url: https://github.com/VNG-Realisatie/Haal-Centraal-BRP-historie-bevragen
  license:
    name: European Union Public License, version 1.2 (EUPL-1.2)
    url: https://eupl.eu/1.2/nl/
tags:
  - name: BRP historie bevragen
    description: Zoeken historische gegevens
paths:
  /verblijfplaatshistorie:
    post:
      description: |
        Raadpleeg de verblijfplaatshistorie van een persoon op de opgegeven peildatum of binnen de opgegeven periode.
        Het meest actuele adres staat bovenaan.

        Zoek met burgerservicenummer
      operationId: Verblijfplaatshistorie
      requestBody:
        content:
          application/json:
            schema:
              $ref: "openapi.yaml#/components/schemas/HistorieQuery"
      responses:
        '200':
          description: |
                        Zoekactie geslaagd
          headers:
            warning:
              $ref: "common.yaml#/components/headers/warning"
            X-Rate-Limit-Limit:
              $ref: "common.yaml#/components/headers/X_Rate_Limit_Limit"
            X-Rate-Limit-Remaining:
              $ref: "common.yaml#/components/headers/X_Rate_Limit_Remaining"
            X-Rate-Limit-Reset:
              $ref: "common.yaml#/components/headers/X_Rate_Limit_Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerblijfplaatshistorieQueryResponse"
        '400':
          $ref: "common.yaml#/components/responses/400"
        '401':
          $ref: "common.yaml#/components/responses/401"
        '403':
          $ref: "common.yaml#/components/responses/403"
        '406':
          $ref: "common.yaml#/components/responses/406"
        '415':
          $ref: 'common.yaml#/components/responses/415'
        '429':
          $ref: "common.yaml#/components/responses/429"
        '500':
          $ref: "common.yaml#/components/responses/500"
        '503':
          $ref: "common.yaml#/components/responses/503"
        'default':
          $ref: "common.yaml#/components/responses/default"
      tags:
        - BRP historie bevragen
# Waar mogelijk wordt verwezen naar componenten uit HC BRP Personen. Dat kan alleen als deze zelf niet ook een $ref bevatten. In dat geval zijn de componenten in de openapi.yaml van HC BRP Historie gekopieerd.

# De vraag is of dit de gewenste aanpak is of dat we alle componenten lokaal binnen de openapi.yaml van HC BRP Historie moeten definiëren.
components:
  schemas:
    VerblijfplaatshistorieQueryResponse:
      type: object
      properties:
        verblijfplaatsen:
          type: array
          items:
            $ref: 'verblijfplaats-voorkomen.yaml#/components/schemas/GbaVerblijfplaatsVoorkomen'
        opschortingBijhouding:
          $ref: 'opschortingbijhouding.yaml#/components/schemas/GbaOpschortingBijhouding'
        geheimhoudingPersoonsgegevens:
          $ref: 'persoon.yaml#/components/schemas/GbaGeheimhoudingPersoonsgegevens'